Work for release 0.23.x

+++ src/main/java/org/killbill/billing/invoice/api/InvoiceItem.java
@@ -79,7 +79,7 @@ public interface InvoiceItem extends Entity {
 
     UUID getLinkedItemId();
 
-    Integer getQuantity();
+    BigDecimal getQuantity();
 
     String getItemDetails();
